Move logger and string-to-array-buffer tests out of test/util/index.test.js
